deployment_hf/requirements.txt
gradio==3.34.0
gradio_client==0.2.6
#onnxruntime==1.15.1 
ohmeow-blurr==1.0.5
torch==1.13.1
transformers==4.33.2
#fastai==2.7.12